ml-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1
scikit-learn==1.6.1
xgboost==2.1.3
pandas==2.2.3
numpy==2.2.1
joblib==1.4.2
python-dotenv==1.0.1
httpx==0.28.1
psycopg2-binary==2.9.10
matplotlib==3.9.3
seaborn==0.13.2
